+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - Added CLOB as STRING alias.
 - Fixed `Float` generator bug that creted numbers bigger than desired DECIMAL precision.
 - Fixed `IMPORT INTO` statement created by `util csv` command.
+- SimpleFaker: cleaned up code